# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Seatsurfing REST API
  description: |
    Seatsurfing provides REST APIs used by the mobile and web app as well as the administrator web interface.

    ## Authorization checks

    Each API endpoint is protected by authorization checks:
    1. **Role check**: Role of the user (Anonymous, User, Space Admin, Org Admin, Super Admin)
    2. **Organization check**: Organization membership

    ## Roles

    - **Anonymous**: Every user allowed, no authorization required
    - **User**: Authorized user who is a member of an organization
    - **Space Admin**: An organization's space administrator (can manage locations, spaces, bookings)
    - **Org Admin**: An organization's full administrator
    - **Super Admin**: Global administrator with access to all organizations and all functions

    ## Authentication

    Authentication is done via JWT Bearer tokens in the `Authorization` header. Service accounts may also use HTTP Basic Authentication.

    ## Session management

    - Access Token Lifetime: 15 minutes
    - Refresh Token Lifetime: 28 days
    - Tokens are signed using RS512

    ## Custom error codes

    Some endpoints return custom error codes via the `X-Error-Code` response header:
    - `1001`: Booking slot conflict
    - `1002`: Location max concurrent bookings reached
    - `1003`: Too many upcoming bookings for user
    - `1004`: Too many days in advance
    - `1005`: Invalid booking duration
...
    - `1010`: Booking subject required
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained via login or refresh endpoints
    BasicAuth:
      type: http
      scheme: basic
      description: Service account authentication using `{orgId}_{email}` as username and the service account password
...
  responses:
...
    Forbidden:
      description: Authorized, but user has no access to requested resource
    NotFound:
      description: Object not found
...
    InternalServerError:
      description: Internal server error
paths:
...
  /booking/:
    get:
      tags: [Bookings]
      summary: Get own upcoming bookings
      description: Returns the authenticated user's upcoming bookings (bookings that haven't ended yet in the location's timezone).
      operationId: getOwnBookings
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of upcoming bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetBookingResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags: [Bookings]
      summary: Create a booking
      description: |
        Creates a new booking. Space Admins can book on behalf of other users by providing `userEmail`.
        Various validation checks are performed (max bookings, max days in advance, slot conflicts, etc.).
      operationId: createBooking
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Slot conflict (space already booked for that time)
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /recurring-booking/:
    post:
      tags: [Recurring Bookings]
      summary: Create a recurring booking
      description: Creates a recurring booking series. Individual bookings are created for each occurrence. Returns the result for each occurrence indicating success or failure.
      operationId: createRecurringBooking
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecurringBookingRequest"
      responses:
        "201":
          description: Recurring booking created
          headers:
            X-Object-ID:
              schema:
                type: string
              description: ID of the recurring booking
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreateRecurringBookingResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "402":
          $ref: "#/components/responses/PaymentRequired"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /location/:
    get:
      tags: [Locations]
      summary: Get all locations
      description: Returns all locations in the user's organization.
      operationId: getAllLocations
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetLocationResponse"
    post:
      tags: [Locations]
      summary: Create a location
      description: Creates a new location in the user's organization. Requires Space Admin role.
      operationId: createLocation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocationRequest"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /location/search:
    post:
      tags: [Locations]
      summary: Search locations
      description: Searches locations with optional attribute filters (e.g. number of free spaces, buddies on site). If no attributes are provided, returns all locations.
      operationId: searchLocations
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchLocationRequest"
      responses:
        "200":
          description: Matching locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetLocationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /location/{locationId}/space/:
    get:
      tags: [Spaces]
      summary: Get all spaces in a location
      description: Returns all spaces in a location with their attributes, approvers, and allowed bookers.
      operationId: getAllSpaces
      security:
        - BearerAuth: []
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of spaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetSpaceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Spaces]
      summary: Create a space
      description: Creates a new space in a location. Requires Space Admin role.
      operationId: createSpace
      security:
        - BearerAuth: []
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpaceRequest"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /space-attribute/:
    get:
      tags: [Space Attributes]
      summary: Get all space attributes
      description: Returns all space attribute definitions for the user's organization.
      operationId: getAllSpaceAttributes
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of space attributes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetSpaceAttributeResponse"
    post:
      tags: [Space Attributes]
      summary: Create a space attribute
      description: Creates a new space attribute definition. Requires Space Admin role.
      operationId: createSpaceAttribute
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpaceAttributeRequest"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /space-type/:
    get:
      tags: [Space Types]
      summary: Get all space types
      description: Returns all seat classifications for the user's organization.
      operationId: getAllSpaceTypes
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of space types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetSpaceTypeResponse"
    post:
      tags: [Space Types]
      summary: Create a space type
      description: Creates a new seat classification and booking rule. Requires Space Admin role.
      operationId: createSpaceType
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpaceTypeRequest"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /organization/:
    get:
      tags: [Organizations]
      summary: Get all organizations
      description: Returns all organizations. Requires Super Admin role.
      operationId: getAllOrganizations
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetOrganizationResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Organizations]
      summary: Create an organization
      description: Creates a new organization. Requires Super Admin role.
      operationId: createOrganization
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /organization/{id}/domain/{domain}:
    post:
      tags: [Organizations]
      summary: Add a domain to an organization
      description: Adds a new domain to an organization. The domain must be verified via DNS TXT record before it becomes active. Requires Org Admin or Super Admin role. Requires the Custom Domains feature.
      operationId: addDomain
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: domain
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "402":
          $ref: "#/components/responses/PaymentRequired"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Organizations]
      summary: Remove a domain from an organization
      description: Removes a domain from an organization. Seatsurfing signup domains (*.seatsurfing.app) cannot be removed. Requires Org Admin or Super Admin role.
      operationId: removeDomain
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: domain
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /organization/{id}/domain/{domain}/primary:
    post:
      tags: [Organizations]
      summary: Set primary domain
      description: Sets a domain as the primary domain for an organization. Requires Org Admin or Super Admin role.
      operationId: setPrimaryDomain
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: domain
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /user/:
    get:
      tags: [Users]
      summary: Get all users
      description: Returns all users of the organization. Supports search via `q` query parameter. Requires Space Admin role (Org Admin for search by keyword).
      operationId: getAllUsers
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search keyword to filter users by email, first name, or last name
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetUserResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Users]
      summary: Create a user
      description: |
        Creates a new user in the organization. Requires Org Admin role. Authentication method is determined by the request:
        - Set `sendInvitation: true` to send an email invitation for the user to set their own password
        - Set `password` to create a user with a specific password
        - Set `authProviderId` to bind the user to an OAuth provider
      operationId: createUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "402":
          $ref: "#/components/responses/PaymentRequired"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: User with this email already exists
...
  /group/:
    get:
      tags: [Groups]
      summary: Get all groups
      description: Returns all groups in the organization. Requires Space Admin role.
      operationId: getAllGroups
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetGroupResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Groups]
      summary: Create a group
      description: Creates a new group in the organization. Requires Org Admin role. Requires the Groups feature.
      operationId: createGroup
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "402":
          $ref: "#/components/responses/PaymentRequired"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /buddy/:
    get:
      tags: [Buddies]
      summary: Get all buddies
      description: Returns all buddies for the authenticated user, including their next booking details.
      operationId: getAllBuddies
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of buddies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetBuddyResponse"
    post:
      tags: [Buddies]
      summary: Add a buddy
      description: Adds a user as a buddy. The buddy must be in the same organization.
      operationId: createBuddy
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBuddyRequest"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /auth-provider/:
    get:
      tags: [Auth Providers]
      summary: Get all auth providers
      description: Returns all authentication providers for the organization. Requires Org Admin role.
      operationId: getAllAuthProviders
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of auth providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetAuthProviderResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Auth Providers]
      summary: Create an auth provider
      description: Creates a new OAuth/OIDC authentication provider. Requires Org Admin role. Requires the Auth Providers feature.
      operationId: createAuthProvider
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAuthProviderRequest"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "402":
          $ref: "#/components/responses/PaymentRequired"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /setting/:
    get:
      tags: [Settings]
      summary: Get all settings
      description: |
        Returns all organization settings visible to the current user.
        Regular users see public settings; Org Admins see additional admin settings;
        Space Admins see admin menu items. Also includes system settings and plugin settings.
      operationId: getAllSettings
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of settings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetSettingsResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [Settings]
      summary: Update all settings
      description: Updates multiple settings at once. Requires Org Admin role.
      operationId: setAllSettings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/GetSettingsResponse"
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Invalid setting name
        "409":
          $ref: "#/components/responses/Conflict"
...
  /preference/:
    get:
      tags: [User Preferences]
      summary: Get all user preferences
      description: Returns all preferences for the authenticated user.
      operationId: getAllPreferences
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of user preferences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetSettingsResponse"
    put:
      tags: [User Preferences]
      summary: Update all user preferences
      description: Updates multiple user preferences at once.
      operationId: setAllPreferences
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/GetSettingsResponse"
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Invalid preference name
...